requirements.txt
python-telegram-bot==20.7
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-dotenv==1.0.0
pydantic==2.5.3
mypy==1.8.0
types-requests==2.31.0 